+++ RenderSystems/GL3Plus/src/OgreGL3PlusRenderSystem.cpp
@@ -333,7 +333,7 @@ namespace Ogre {
 
         GLint maxRes2d;
         OGRE_CHECK_GL_ERROR(glGetIntegerv(GL_MAX_TEXTURE_SIZE, &maxRes2d));
-        rsc->setMaximumResolution2D(static_cast<ushort>(maxRes2d));
+        rsc->setMaximumResolution2D(static_cast<uint32>(maxRes2d));
 
         // Point size
         GLfloat psRange[2] = {0.0, 0.0};
